refactor(portalnet): split out overlay_service into submodules
