+++ trin-beacon/src/jsonrpc.rs
@@ -14,7 +14,7 @@ use ethportal_api::{
     utils::bytes::hex_encode,
     BeaconContentKey, BeaconContentValue, OverlayContentKey, RawContentKey,
 };
-use portalnet::overlay_service::OverlayRequestError;
+use portalnet::overlay::errors::OverlayRequestError;
 use serde_json::{json, Value};
 use tokio::sync::{mpsc, Mutex, RwLock};
 use tracing::error;
